Move main Visual type declaration into react-visual
