+++ packages/next/src/index.ts
@@ -1,7 +1,3 @@
 import NextVisual from './NextVisual'
 export default NextVisual
-export {
-  NextVisualProps,
-  ObjectFit,
-  ObjectFitOption,
-} from './types/nextVisualTypes'
+export { NextVisualProps } from './types/nextVisualTypes'
