+++ packages/next/src/types/nextVisualTypes.ts
@@ -1,34 +1,3 @@
-import type { CSSProperties } from 'react'
+import type { ReactVisualTypes } from '@react-visual/react'
 
-export type NextVisualProps = {
-
-  image?: string
-  video?: string
-  placeholderData?: string
-
-  expand?: boolean
-  aspect?: number // An explict aspect ratio
-  width?: number | string
-  height?: number | string
-  fit?: ObjectFitOption | ObjectFit
-  position?: string
-
-  priority?: boolean
-  sizes?: string
-  imageLoader?: Function
-
-  paused?: boolean
-
-  alt: string
-
-  className?: string
-  style?: CSSProperties
-}
-
-export type ObjectFitOption = 'cover' | 'contain'
-
-// Deprecated
-export enum ObjectFit {
-  Cover = 'cover',
-  Contain = 'contain',
-}
+export type NextVisualProps = ReactVisualTypes
