+++ packages/react/src/index.ts
@@ -2,3 +2,8 @@ import LazyVideo from './LazyVideo'
 import VisualWrapper from './VisualWrapper'
 import { collectDataAttributes } from './lib/attributes'
 export { LazyVideo, VisualWrapper, collectDataAttributes }
+export {
+  ReactVisualTypes,
+  ObjectFit,
+  ObjectFitOption
+} from './types/reactVisualTypes'
